+++ lib/sec_id/cfi.rb
@@ -189,12 +189,6 @@ module SecId
       @attr2 = cfi_parts[:attr2]
       @attr3 = cfi_parts[:attr3]
       @attr4 = cfi_parts[:attr4]
-      @check_digit = nil
-    end
-
-    # @return [Boolean] always false - CFI has no check digit
-    def has_check_digit?
-      false
     end
 
     # Validates format including category and group codes.