+++ lib/sec_id/occ.rb
@@ -97,12 +97,6 @@ module SecId
       @date_str = symbol_parts[:date]
       @type = symbol_parts[:type]
       @strike_mills = symbol_parts[:strike_mills]
-      @check_digit = nil
-    end
-
-    # @return [Boolean] always false
-    def has_check_digit?
-      false
     end
 
     # Normalizes the OCC symbol to standard format with 6-char padded underlying and 8-digit strike.