+++ spec/sec_id/wkn_spec.rb
@@ -11,15 +11,16 @@ RSpec.describe SecId::WKN do
 
     it 'parses WKN correctly' do
       expect(wkn.identifier).to eq('514000')
-      expect(wkn.check_digit).to be_nil
     end
 
     it 'returns the full number' do
       expect(wkn.full_number).to eq('514000')
     end
 
-    it 'has no check digit' do
-      expect(wkn.has_check_digit?).to be(false)
+    describe '#to_s' do
+      it 'returns the identifier' do
+        expect(wkn.to_s).to eq('514000')
+      end
     end
   end
 
@@ -28,7 +29,6 @@ RSpec.describe SecId::WKN do
 
     it 'parses WKN correctly' do
       expect(wkn.identifier).to eq('CBK100')
-      expect(wkn.check_digit).to be_nil
     end
   end